+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.7.8 [unreleased]
 
+- Added `boost_by` and `boost_where` options
 - Added `select` option for `load: false`
 
 ## 0.7.7